// src/enums/TileType.ts
export enum TileType {
    Grass = 1,
    WaterCornerOne = 2,
    WaterCornerTwo = 3,
    WaterCornerThree = 4,
    WaterCornerFour = 5,
    WaterVOne = 6,
    WaterHOne = 7,
    WaterHTwo = 8,
    WaterVTwo = 9,
    Goblin = 10,
    Tree = 11,
    Tower = 12,
    Hut = 13,
}

export enum AudioType {
    Message = 'drop.ogg',
    Construct = 'construct.ogg'
}